sony: loire: Fix iddd service
iddd service uses /idd partition which is ...by-name/diag block device.
We are using apps_log for /misc which is required for recovery
and it breaks /rca requirements

So this patch creates the symlink /idd/rca -> /rca
Since /rca has not been used for OmniROM then there is no
problem to set this symlink for now to avoid logspam.

Now iddd service is ready to serve the world.
12-19 15:32:44.832   435   435 I iddd    : Ready to serve the world.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: If5c1f63f89eeb2e4920ef41c289dc69f8c02ba03